Crate problem021

Problem 21 (Amicable numbers)

Solution detail

We will check each number up to 10000 in turn to see if it is amicable. To do this, it will be necessary to be able to compute d(n) for given n. Then, to check if n is amicable, we just check that d(d(n)) = n and that d(n) ≠ n.

To calculate the sum of the divisors of a given number, it will be necessary to be able to fully factorise that number, which requires having sieved primes up to the square root of that number. It would be good to know, in advance, the largest number that we have to factorise, so that we can presieve the primes just once. The largest number we might see is the maximum value of d(n) for n ≤ 10000.

Thankfully, under the assumption of the Riemann Hypothesis, Robin's Theorem gives us an upper bound on this maximum. It says that for all n > 5040, d(n) < e^γ * log(log(n)) * n. Since for n ≤ 5040, d(n) < 4n, this means that we need the highest number we might need to be able to factorise is:

max(4, e^γ * log(log(10000))) * 10000

Functions

answer
Solve the problem, returning the answer as a String

is_amicable
Check whether n is amicable.

main

solve
Find the sum of the amicable numbers below the given limit.